+++ History.md
@@ -1,3 +1,11 @@
+5.5.9 / 2019-05-16
+==================
+ * fix(query): skip schema setters when casting $regexp $options #7802 [Fonger](https://github.com/Fonger)
+ * fix(populate): don't skip populating doc array properties whose name conflicts with an array method #7782
+ * fix(populate): make populated virtual return undefined if not populated #7795
+ * fix(schema): handle custom setters in arrays of document arrays #7804 [Fonger](https://github.com/Fonger)
+ * docs(tutorials): add query casting tutorial #7789
+
 5.5.8 / 2019-05-13
 ==================
  * fix(document): run pre save hooks on nested child schemas #7792